Add tabs to QuestActivity. Create TimerFragment
