+++ app/src/main/res/layout/activity_quest.xml
@@ -6,87 +6,34 @@
     android:layout_height="match_parent"
     android:background="@color/md_blue_800">
 
-    <RelativeLayout
-        android:id="@+id/quest_details_container"
+    <android.support.design.widget.AppBarLayout
+        android:id="@+id/appbar"
         android:layout_width="match_parent"
-        android:layout_height="match_parent">
+        android:layout_height="wrap_content"
+        android:background="?attr/colorPrimary"
+        app:elevation="0dp"
+        android:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar">
 
-        <TextView
-            android:id="@+id/quest_details_name"
-            style="@style/TextAppearance.AppCompat.Headline"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_above="@+id/quest_details_progress"
-            android:layout_centerHorizontal="true"
-            android:layout_margin="32dp"
-            android:ellipsize="end"
-            android:fontFamily="sans-serif"
-            android:gravity="center"
-            android:maxLines="2"
-            android:textColor="@color/md_light_text_87" />
-
-        <ProgressBar
-            android:id="@+id/quest_details_progress"
-            style="?android:attr/progressBarStyleHorizontal"
-            android:layout_width="320dp"
-            android:layout_height="320dp"
-            android:layout_centerInParent="true"
-            android:indeterminate="false"
-            android:max="100"
-            android:progress="0"
-            android:progressDrawable="@drawable/timer_progress"
-            android:secondaryProgress="100" />
-
-        <Chronometer
-            android:id="@+id/quest_details_time"
-            style="@style/TextAppearance.AppCompat.Display4"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_centerInParent="true"
-            android:fontFamily="sans-serif-thin"
-            android:textColor="@color/md_light_text_87"
-            android:textSize="72sp" />
-
-    </RelativeLayout>
-
-    <RelativeLayout
+    <android.support.v7.widget.Toolbar
+        android:id="@+id/toolbar"
         android:layout_width="match_parent"
-        android:layout_height="72dp"
-        android:layout_marginEnd="32dp"
-        android:layout_marginStart="32dp"
-        android:layout_marginTop="16dp"
-        app:layout_anchor="@id/quest_details_container"
-        app:layout_anchorGravity="bottom">
+        android:layout_height="?attr/actionBarSize"
+        app:layout_collapseMode="pin"
+        app:popupTheme="@style/ThemeOverlay.AppCompat.Light" />
 
-
-        <android.support.design.widget.FloatingActionButton
-            android:id="@+id/quest_details_timer"
-            android:layout_width="wrap_content"
+        <android.support.design.widget.TabLayout
+            android:id="@+id/tabs"
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_centerInParent="true"
-            android:clickable="true"
-            android:scaleType="center"
-            android:src="@drawable/ic_play_arrow_white_32dp" />
+            app:tabMode="fixed"
+            app:tabGravity="fill"
+            app:tabIndicatorColor="@android:color/transparent" />
+    </android.support.design.widget.AppBarLayout>
 
-        <ImageButton
-            android:id="@+id/quest_details_edit"
-            android:layout_width="36dp"
-            android:layout_height="36dp"
-            android:layout_alignParentStart="true"
-            android:layout_centerVertical="true"
-            android:background="@android:color/transparent"
-            android:scaleType="fitXY"
-            android:src="@drawable/ic_create_white_24dp" />
-
-        <ImageButton
-            android:id="@+id/quest_details_done"
-            android:layout_width="36dp"
-            android:layout_height="36dp"
-            android:layout_alignParentEnd="true"
-            android:layout_centerVertical="true"
-            android:background="@android:color/transparent"
-            android:scaleType="fitXY"
-            android:src="@drawable/ic_done_white_24dp" />
-    </RelativeLayout>
+    <android.support.v4.view.ViewPager
+        android:id="@+id/view_pager"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        app:layout_behavior="@string/appbar_scrolling_view_behavior"  />
 
 </android.support.design.widget.CoordinatorLayout>
